distributed training with Google AI Platform - part 1
This CR includes:
- Refactoring of code under train/ which allows us to separate out code
used for training not dependant on the rest of the code base. This makes
it easy to understand the context used for training.
- Add a Dockerfile to build a custom image used for Google AI Platform,
as well as a script to drive training and post-training inference.
- Write test to some of the training flow - however, manual tests should
still be ran to test training locally and training with the GCP script,
just to be sure.
- Some lint fixes

Note: It turns out we can do both train _and_ batch inference with the
Google AI Platform. Hurray for killing two birds with one stone!